# project/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "dynamic yolo",
    "version": "1.0.0",
    "title": "Dynamic yolo",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/projects": {
      "post": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Create a new project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/project"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created new user"
          },
          "409": {
            "description": "User has been created"
          }
        }
      },
      "get": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Get all projects",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/projects/{project_id}/upload": {
      "post": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Upload files to project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "description": "Project id",
            "schema": {
              "$ref": "#/components/schemas/project_id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/upload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Get project info by id",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/project_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/project"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/settings": {
      "put": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Change project settings",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/project_id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/project_settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project settings updated"
          },
          "400": {
            "description": "Invalid project ID or settings"
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "get": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Get project settings",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/project_id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_project_settings"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/images": {
      "get": {
        "tags": [
          "Project Requests"
        ],
        "summary": "Get images for a project",
        "description": "Get a paginated list of images for a project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "ID of the project to get images for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of images to be returned per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "page_nr",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of images for the project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/image"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/models": {
      "get": {
        "tags": [
          "Project Requests"
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "ID of the project to get models for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of models to be returned per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "page_nr",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "summary": "Get all models in a project",
        "responses": {
          "200": {
            "description": "List of models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/simple_model_info"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/models/{model_id}": {
      "get": {
        "tags": [
          "Models Requests"
        ],
        "summary": "Get model info",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "Model id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/model_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/model_info"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/annotations": {
      "get": {
        "tags": [
          "Project Requests"
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "ID of the project to get annotations for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of annotations to be returned per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "page_nr",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "summary": "Get all annotations of a project",
        "responses": {
          "200": {
            "description": "List of annotations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/simple_annotation_info"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/results": {
      "get": {
        "tags": [
          "Results Requests"
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Specify the project of which results you want to see",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "description": "Model ID of the result",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "subset_name",
            "in": "query",
            "description": "Name of the subset",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "test",
                "train"
              ]
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to be returned per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "page_nr",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "summary": "Get all the results",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/simple_model_results"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid project"
          }
        }
      }
    },
    "/results/{result_id}": {
      "get": {
        "tags": [
          "Results Requests"
        ],
        "summary": "Get detailed info about trained model",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "description": "Result ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Results info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/detailed_model_results"
                }
              }
            }
          },
          "404": {
            "description": "Result not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "project_id": {
        "type": "integer"
      },
...
      "project": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "format": "string"
          },
          "max_class_nr": {
            "type": "integer"
          },
          "img_size": {
            "type": "integer"
          },
          "initial_model": {
            "type": "string",
            "format": "string"
          }
        }
      },
      "upload": {
        "type": "object",
        "properties": {
          "uploader_name": {
            "type": "string",
            "format": "string"
          },
          "split": {
            "type": "string",
            "format": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
      "project_settings": {
        "type": "object",
        "properties": {
          "epochs": {
            "type": "integer",
            "description": "Number of training epochs"
          },
          "batch_size": {
            "type": "integer",
            "description": "Batch size for training"
          },
          "train_test_ratio": {
            "type": "number",
            "description": "Training and testing data split ratio"
          },
          "pretest_size_difference": {
            "type": "number"
          },
          "maximum_auto_train_number": {
            "type": "number"
          },
          "skip_pretesting": {
            "type": "boolean",
            "default": false
          },
          "minimal_map_50_threshold": {
            "type": "number",
            "description": "Minimal map threshold"
          },
          "minimal_map_50_95_threshold": {
            "type": "number"
          },
          "minimal_precision_threshold": {
            "type": "number"
          },
          "minimal_recall_threshold": {
            "type": "number"
          },
          "min_confidence_threshold": {
            "type": "number",
            "description": "Minimum confidence threshold for detection"
          },
          "min_iou_threshold": {
            "type": "number",
            "description": "Minimum intersection over union threshold for detection"
          }
        }
      },
      "get_project_settings": {
        "type": "object",
        "properties": {
          "max_class_nr": {
            "type": "integer"
          },
          "epochs": {
            "type": "integer"
          },
          "batch_size": {
            "type": "integer"
          },
          "img_size": {
            "type": "integer"
          },
          "initial_model": {
            "type": "string",
            "format": "string"
          },
          "train_test_ratio": {
            "type": "integer"
          },
          "pretest_size_difference_threshold": {
            "type": "number"
          },
          "maximum_auto_train_number": {
            "type": "integer"
          },
          "skip_pretesting": {
            "type": "boolean"
          },
          "minimal_map_50_threshold": {
            "type": "integer"
          },
          "minimal_map_50_95_threshold": {
            "type": "integer"
          },
          "minimal_precision_threshold": {
            "type": "integer"
          },
          "minimal_recall_threshold": {
            "type": "integer"
          },
          "min_confidence_threshold": {
            "type": "integer"
          },
          "min_iou_threshold": {
            "type": "integer"
          }
        }
      },
      "simple_model_info": {
        "type": "object",
        "properties": {
          "model_status_name": {
            "type": "string",
            "description": "Status of the model"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the model"
          },
          "added": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time the model was added"
          },
          "project_id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the project model is associated with"
          }
        }
      },
      "model_id": {
        "type": "integer"
      },
      "simple_annotation_info": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID of the annotation"
          },
          "project_id": {
            "type": "integer",
            "description": "ID of the project"
          },
          "annotator_name": {
            "type": "string",
            "description": "Annotator of the annotation"
          },
          "image_id": {
            "type": "integer",
            "description": "ID of the image"
          }
        }
      },
      "image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "annotations": {
            "type": "array",
            "items": {
              "type": "integer",
              "description": "ID-s of annotations"
            }
          },
          "models": {
            "type": "array",
            "items": {
              "type": "integer",
              "description": "ID-s of models"
            }
          },
          "subset_name": {
            "type": "string",
            "format": "string"
          }
        }
      },
      "model_info": {
        "type": "object",
        "properties": {
          "parent_model_id": {
            "type": "integer",
            "format": "int64"
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "total_epochs": {
            "type": "integer",
            "format": "int32"
          },
          "epochs": {
            "type": "integer",
            "format": "int32"
          },
          "model_status_name": {
            "type": "string"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          },
          "model_results": {
            "type": "array",
            "items": {
              "type": "integer",
              "description": "ID-s of results"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "integer",
              "description": "ID-s of images"
            }
          }
        },
        "required": [
          "id",
          "added",
          "total_epochs",
          "epochs",
          "model_status_id",
          "project_id"
        ]
      },
...
      "simple_model_results": {
        "type": "object",
        "properties": {
          "result_id": {
            "type": "integer"
          },
          "subset_name": {
            "type": "string"
          },
          "model_id": {
            "type": "integer"
          },
          "metric_precision": {
            "type": "number",
            "format": "float"
          },
          "metric_recall": {
            "type": "number",
            "format": "float"
          },
          "metric_map_50": {
            "type": "number",
            "format": "float"
          },
          "metric_map_50_95": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "detailed_model_results": {
        "type": "object",
        "properties": {
          "result_id": {
            "type": "integer",
            "format": "int64"
          },
          "model_id": {
            "type": "integer",
            "format": "int64"
          },
          "project_id": {
            "type": "integer"
          },
          "subset_name": {
            "type": "string"
          },
          "epoch": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "metric_precision": {
            "type": "number",
            "format": "float"
          },
          "metric_recall": {
            "type": "number",
            "format": "float"
          },
          "metric_map_50": {
            "type": "number",
            "format": "float"
          },
          "metric_map_50_95": {
            "type": "number",
            "format": "float"
          },
          "val_box_loss": {
            "type": "number",
            "format": "float"
          },
          "val_obj_loss": {
            "type": "number",
            "format": "float"
          },
          "val_cls_loss": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "model_id",
          "subset_id",
          "metric_precision",
          "metric_recall",
          "metric_map_50",
          "metric_map_50_95",
          "val_box_loss",
          "val_obj_loss",
          "val_cls_loss"
        ]
      }
    }
  }
}